refactor: guard OCP scheme registrations with distro build tags
Prometheus Operator monitoring APIs, OpenShift Route, and Istio v1
scheme registrations were wired unconditionally, pulling OCP-only
dependencies into upstream builds. This makes the upstream binary carry
dead scheme registrations and pulls in imports that don't belong there.

Uses the established companion file pattern (//go:build distro /
!distro) to gate these registrations, keeping the upstream build surface
clean while preserving full functionality in distro builds.
