+++ cmd/manager/main_schemes_default.go
@@ -1,5 +1,7 @@
+//go:build !distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,22 +16,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testing
+package main
 
-import (
-	"path/filepath"
+import "k8s.io/apimachinery/pkg/runtime"
 
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
-)
-
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemeFuncs := []AddToSchemeFunc{kservescheme.AddAll}
-	schemeFuncs = append(schemeFuncs, additionalTestSchemes()...)
-	schemes := WithScheme(schemeFuncs...)
-
-	return Configure(append(options, testCRDs, schemes)...)
+// registerDistroSchemes is a no-op in upstream builds.
+func registerDistroSchemes(_ *runtime.Scheme) error {
+	return nil
 }
