+++ pkg/scheme/register_ocp.go
@@ -1,5 +1,7 @@
+//go:build distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,22 +16,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testing
+package scheme
 
 import (
-	"path/filepath"
-
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
+	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 )
 
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemeFuncs := []AddToSchemeFunc{kservescheme.AddAll}
-	schemeFuncs = append(schemeFuncs, additionalTestSchemes()...)
-	schemes := WithScheme(schemeFuncs...)
+func init() {
+	distroSchemes = append(distroSchemes, AddMonitoringAPIs)
+}
 
-	return Configure(append(options, testCRDs, schemes)...)
+// AddMonitoringAPIs registers Prometheus Operator monitoring APIs (PodMonitor, ServiceMonitor).
+func AddMonitoringAPIs(s *runtime.Scheme) error {
+	return addAll(s, monitoringv1.AddToScheme)
 }
