+++ pkg/testing/ctrl_ocp.go
@@ -1,5 +1,7 @@
+//go:build distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,19 +19,15 @@ limitations under the License.
 package testing
 
 import (
-	"path/filepath"
-
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
+	routev1 "github.com/openshift/api/route/v1"
+	istioclientv1 "istio.io/client-go/pkg/apis/networking/v1"
 )
 
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemeFuncs := []AddToSchemeFunc{kservescheme.AddAll}
-	schemeFuncs = append(schemeFuncs, additionalTestSchemes()...)
-	schemes := WithScheme(schemeFuncs...)
-
-	return Configure(append(options, testCRDs, schemes)...)
+// additionalTestSchemes returns OCP-specific scheme registration functions
+// for the shared envtest setup (OpenShift Routes, Istio v1).
+func additionalTestSchemes() []AddToSchemeFunc {
+	return []AddToSchemeFunc{
+		routev1.AddToScheme,
+		istioclientv1.AddToScheme,
+	}
 }
