+++ pkg/testing/envtest_setup_default.go
@@ -1,5 +1,7 @@
+//go:build !distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,20 +18,7 @@ limitations under the License.
 
 package testing
 
-import (
-	"path/filepath"
+import "k8s.io/apimachinery/pkg/runtime"
 
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
-)
-
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemeFuncs := []AddToSchemeFunc{kservescheme.AddAll}
-	schemeFuncs = append(schemeFuncs, additionalTestSchemes()...)
-	schemes := WithScheme(schemeFuncs...)
-
-	return Configure(append(options, testCRDs, schemes)...)
-}
+// registerLegacyDistroSchemes is a no-op in upstream builds.
+func registerLegacyDistroSchemes(_ *runtime.Scheme) {}
