+++ pkg/testing/envtest_setup_ocp.go
@@ -1,5 +1,7 @@
+//go:build distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,19 +19,14 @@ limitations under the License.
 package testing
 
 import (
-	"path/filepath"
-
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
+	istioclientv1 "istio.io/client-go/pkg/apis/networking/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 )
 
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemeFuncs := []AddToSchemeFunc{kservescheme.AddAll}
-	schemeFuncs = append(schemeFuncs, additionalTestSchemes()...)
-	schemes := WithScheme(schemeFuncs...)
-
-	return Configure(append(options, testCRDs, schemes)...)
+// registerLegacyDistroSchemes registers OCP-specific schemes for the legacy
+// SetupEnvTest path (Istio v1 networking APIs).
+func registerLegacyDistroSchemes(s *runtime.Scheme) {
+	if err := istioclientv1.SchemeBuilder.AddToScheme(s); err != nil {
+		log.Error(err, "Failed to add istio v1 scheme")
+	}
 }
